+++ callback.js
@@ -41,7 +41,10 @@ module.exports.createConnection = function createConnection(opts) {
 
 exports.createPool = function createPool(opts) {
   const options = new PoolOptions(opts);
-  return new PoolCallback(options);
+  const pool = new PoolCallback(options);
+  // adding a default error handler to avoid exiting application on connection error.
+  pool.on('error', (err) => {});
+  return pool;
 };
 
 exports.createPoolCluster = function createPoolCluster(opts) {