+++ test/integration/test-connection-opts.js
@@ -2,7 +2,6 @@
 
 const base = require('../base.js');
 const { assert } = require('chai');
-const { isXpand } = require('../base');
 
 describe('connection option', () => {
   it('with undefined collation', function (done) {
@@ -50,6 +49,9 @@ describe('connection option', () => {
   });
 
   it('timezone Z', async function () {
+    // node.js before v13 doesn't permit to set TZ value repeatedly
+    if (parseInt(process.versions.node.split('.')[0]) <= 12) this.skip();
+
     const defaultTz = process.env.TZ;
     const conn = await base.createConnection({ timezone: 'Z' });
     conn.query("SET SESSION time_zone = '+01:00'");
@@ -67,18 +69,18 @@ describe('connection option', () => {
   });
 
   it('timezone +0h', async function () {
-    const conn = await base.createConnection({ timezone: '+00:00' });
+    const conn = await base.createConnection({ timezone: '+00:00', debug: true });
 
     let d = new Date('2000-01-01T00:00:00Z');
-    let res = await conn.query('SELECT UNIX_TIMESTAMP(?) tt', [d]);
+    let res = await conn.query("SELECT UNIX_TIMESTAMP('2000-01-01T00:00:00') tt", [d]);
     assert.equal(Number(res[0].tt), d.getTime() / 1000);
-    res = await conn.query("SELECT TIMESTAMP('2003-12-31 12:00:00') tt1, FROM_UNIXTIME(UNIX_TIMESTAMP(?)) tt2", [d]);
-    assert.deepEqual(res[0].tt1, new Date('2003-12-31T12:00:00Z'));
-    assert.deepEqual(res[0].tt2, d);
     conn.end();
   });
 
   it('timezone +10h00', async function () {
+    // node.js before v13 doesn't permit to set TZ value repeatedly
+    if (parseInt(process.versions.node.split('.')[0]) <= 12) this.skip();
+
     const defaultTz = process.env.TZ;
     const conn = await base.createConnection({ timezone: '+10:00' });
     process.env.TZ = 'Etc/GMT-10';
@@ -94,6 +96,9 @@ describe('connection option', () => {
   });
 
   it('timezone Etc/GMT-10', async function () {
+    // node.js before v13 doesn't permit to set TZ value repeatedly
+    if (parseInt(process.versions.node.split('.')[0]) <= 12) this.skip();
+
     const defaultTz = process.env.TZ;
     const conn = await base.createConnection({ timezone: 'Etc/GMT-10' });
     process.env.TZ = 'Etc/GMT-10';
@@ -109,6 +114,9 @@ describe('connection option', () => {
   });
 
   it('timezone GMT+10', async function () {
+    // node.js before v13 doesn't permit to set TZ value repeatedly
+    if (parseInt(process.versions.node.split('.')[0]) <= 12) this.skip();
+
     const defaultTz = process.env.TZ;
     const conn = await base.createConnection({ timezone: 'GMT+10' });
     process.env.TZ = 'Etc/GMT-10';
@@ -139,6 +147,9 @@ describe('connection option', () => {
   });
 
   it('Server with different tz', async function () {
+    // node.js before v13 doesn't permit to set TZ value repeatedly
+    if (parseInt(process.versions.node.split('.')[0]) <= 12) this.skip();
+
     if (process.env.srv === 'maxscale' || process.env.srv === 'skysql' || process.env.srv === 'skysql-ha') this.skip();
     //MySQL 5.5 doesn't have milliseconds
     if (!shareConn.info.isMariaDB() && !shareConn.info.hasMinVersion(5, 6, 0)) this.skip();
@@ -343,7 +354,10 @@ describe('connection option', () => {
               done(new Error('must have thrown error'));
             })
             .catch((err) => {
-              assert.isTrue(err.message.includes('Cannot use timeout for MySQL server'));
+              assert.isTrue(
+                err.message.includes('Cannot use timeout for MySQL server') ||
+                  err.message.includes('Cannot use timeout for xpand/MariaDB')
+              );
               assert.equal(err.errno, 45038);
               assert.equal(err.sqlState, 'HY000');
               assert.equal(err.code, 'ER_TIMEOUT_NOT_SUPPORTED');